+++ js/controllers/disks/show.js
@@ -3,16 +3,18 @@ App.DisksShowController = Ember.Controller.extend({
 		deleteDisk: function() {
 			var that = this;
 
-      		var successHandler = function(model) {
+			var successHandler = function(model) {
 				that.transitionToRoute('archives.index');
-     		 }
+			}
 
-	      	var errorHandler = function(reason) {
-       			that.set('hasErrors', true);
-	        	return false;
-	      	}
-
-			that.get('model').destroyRecord().then(successHandler, errorHandler);
+			var errorHandler = function(reason) {
+				that.set('hasErrors', true);
+				that.get('model').rollback();
+				return false;
+			}
+			if (confirm("Delete")){
+				that.get('model').destroyRecord().then(successHandler, errorHandler);
+			}
 		},
 		openContentBrowser: function(diskImage) {
 			window.open('/tree/#'+diskImage.get('id'), '_blank');
